Clean up Tool message display
