+++ packages/frontend/editor-ui/src/features/agents/components/AgentChatEmptyState.vue
@@ -1,0 +1,38 @@
+<script setup lang="ts">
+import { computed } from 'vue';
+import { N8nIcon, N8nText } from '@n8n/design-system';
+
+const props = defineProps<{
+	endpoint: 'build' | 'chat';
+}>();
+
+const icon = computed(() => (props.endpoint === 'build' ? 'wand-sparkles' : 'message-square'));
+const title = computed(() =>
+	props.endpoint === 'build' ? 'Build your agent' : 'Chat with your agent',
+);
+const subtitle = computed(() =>
+	props.endpoint === 'build'
+		? 'Describe what you want your agent to do'
+		: 'Send a message to start a conversation',
+);
+</script>
+
+<template>
+	<div :class="$style.emptyState">
+		<N8nIcon :icon="icon" :size="32" />
+		<N8nText tag="p" bold>{{ title }}</N8nText>
+		<N8nText size="small" color="text-light">{{ subtitle }}</N8nText>
+	</div>
+</template>
+
+<style module>
+.emptyState {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	height: 100%;
+	gap: var(--spacing--3xs);
+	color: var(--color--text--tint-2);
+}
+</style>
