+++ packages/frontend/editor-ui/src/features/agents/components/AgentChatPanel.vue
@@ -1,11 +1,10 @@
 <script setup lang="ts">
-import { ref, reactive, computed, onMounted, nextTick, useTemplateRef, watch } from 'vue';
-import { N8nIcon, N8nText } from '@n8n/design-system';
-import { useRootStore } from '@n8n/stores/useRootStore';
-import { getBuilderMessages, clearBuilderMessages } from '../composables/useAgentApi';
+import { ref, toRef, watch, onMounted } from 'vue';
+import { N8nIcon } from '@n8n/design-system';
 import ChatInputBase from '@/features/ai/shared/components/ChatInputBase.vue';
-import ChatMarkdownChunk from '@/features/ai/chatHub/components/ChatMarkdownChunk.vue';
-import ChatTypingIndicator from '@/features/ai/chatHub/components/ChatTypingIndicator.vue';
+import { useAgentChatStream } from '../composables/useAgentChatStream';
+import AgentChatEmptyState from './AgentChatEmptyState.vue';
+import AgentChatMessageList from './AgentChatMessageList.vue';
 
 const props = withDefaults(
 	defineProps<{
@@ -31,294 +30,40 @@ const emit = defineEmits<{
 	'update:streaming': [streaming: boolean];
 }>();
 
-const rootStore = useRootStore();
-
-interface ChatMessage {
-	id: string;
-	role: 'user' | 'assistant';
-	content: string;
-	thinking?: string;
-	toolCalls?: Array<{ tool: string; input?: unknown; output?: unknown }>;
-	status?: 'streaming' | 'success' | 'error';
-}
-
-const messages = ref<ChatMessage[]>([]);
 const inputText = ref('');
-const isStreaming = ref(false);
-const abortController = ref<AbortController | null>(null);
-const builderHistoryLoaded = ref(false);
-const scrollRef = useTemplateRef<HTMLDivElement>('scrollRef');
 
-const messagingState = computed<'idle' | 'waitingFirstChunk' | 'receiving'>(() => {
-	if (!isStreaming.value) return 'idle';
-	const lastMsg = messages.value[messages.value.length - 1];
-	if (!lastMsg || lastMsg.role === 'user') return 'waitingFirstChunk';
-	return 'receiving';
+const {
+	messages,
+	isStreaming,
+	messagingState,
+	loadBuilderHistory,
+	clearHistory,
+	sendMessage,
+	stopGenerating,
+} = useAgentChatStream({
+	projectId: toRef(props, 'projectId'),
+	agentId: toRef(props, 'agentId'),
+	endpoint: toRef(props, 'endpoint'),
+	onCodeUpdated: () => emit('codeUpdated'),
+	onCodeDelta: (d) => emit('codeDelta', d),
+	onConfigUpdated: () => emit('configUpdated'),
 });
 
 watch(isStreaming, (v) => emit('update:streaming', v));
 
-function scrollToBottom() {
-	void nextTick(() => {
-		if (scrollRef.value) {
-			scrollRef.value.scrollTop = scrollRef.value.scrollHeight;
-		}
-	});
-}
-
-/** Convert persisted agent messages into the frontend ChatMessage format. */
-function convertDbMessages(dbMessages: unknown[]): ChatMessage[] {
-	const result: ChatMessage[] = [];
-	for (const raw of dbMessages) {
-		const msg = raw as {
-			id?: string;
-			role?: string;
-			content?: Array<{
-				type: string;
-				text?: string;
-				toolName?: string;
-				input?: unknown;
-				result?: unknown;
-			}>;
-		};
-		if (!msg.role || !Array.isArray(msg.content)) continue;
-		if (msg.role !== 'user' && msg.role !== 'assistant') continue;
-
-		let text = '';
-		let thinking = '';
-		const toolCalls: Array<{ tool: string; input?: unknown; output?: unknown }> = [];
-
-		for (const part of msg.content) {
-			if (part.type === 'text' && part.text) {
-				text += part.text;
-			} else if (part.type === 'reasoning' && part.text) {
-				thinking += part.text;
-			} else if (part.type === 'tool-call' && part.toolName) {
-				toolCalls.push({ tool: part.toolName, input: part.input });
-			} else if (part.type === 'tool-result' && part.toolName) {
-				const existing = toolCalls.find((t) => t.tool === part.toolName && t.output === undefined);
-				if (existing) {
-					existing.output = part.result;
-				}
-			}
-		}
-
-		result.push({
-			id: msg.id ?? crypto.randomUUID(),
-			role: msg.role,
-			content: text,
-			thinking: thinking || undefined,
-			toolCalls: toolCalls.length > 0 ? toolCalls : undefined,
-		});
-	}
-	return result;
-}
-
-async function loadBuilderHistory() {
-	if (builderHistoryLoaded.value) return;
-	try {
-		const dbMessages = await getBuilderMessages(
-			rootStore.restApiContext,
-			props.projectId,
-			props.agentId,
-		);
-		if (dbMessages.length > 0) {
-			messages.value = convertDbMessages(dbMessages);
-		}
-	} catch {
-		// Silently ignore — just start with empty chat
-	} finally {
-		builderHistoryLoaded.value = true;
-	}
-}
-
 async function onClearHistory() {
-	if (props.endpoint !== 'build') {
-		messages.value = [];
-		return;
-	}
-	try {
-		await clearBuilderMessages(rootStore.restApiContext, props.projectId, props.agentId);
-		messages.value = [];
-	} catch {
-		// ignore
-	}
+	await clearHistory();
 }
 
-async function streamFromEndpoint(endpoint: 'build' | 'chat', message: string) {
-	isStreaming.value = true;
-	let builderMutated = false;
-	const assistantMsg = reactive<ChatMessage>({
-		id: crypto.randomUUID(),
-		role: 'assistant',
-		content: '',
-		thinking: '',
-		toolCalls: [],
-		status: 'streaming',
-	});
-	let msgAdded = false;
-
-	const ensureMsg = () => {
-		if (!msgAdded) {
-			messages.value.push(assistantMsg);
-			msgAdded = true;
-			scrollToBottom();
-		}
-	};
-
-	const controller = new AbortController();
-	abortController.value = controller;
-
-	try {
-		const { baseUrl } = rootStore.restApiContext;
-		const browserId = localStorage.getItem('n8n-browserId') ?? '';
-		const url = `${baseUrl}/projects/${props.projectId}/agents/v2/${props.agentId}/${endpoint}`;
-		const response = await fetch(url, {
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-				'browser-id': browserId,
-			},
-			credentials: 'include',
-			body: JSON.stringify({ message }),
-			signal: controller.signal,
-		});
-
-		if (!response.ok || !response.body) {
-			assistantMsg.content = `Error: ${response.statusText || 'Failed to reach agent'}`;
-			assistantMsg.status = 'error';
-			messages.value.push(assistantMsg);
-			return;
-		}
-
-		const reader = response.body.getReader();
-		const decoder = new TextDecoder();
-		let buffer = '';
-
-		while (true) {
-			const { done, value } = await reader.read();
-			if (done) break;
-
-			buffer += decoder.decode(value, { stream: true });
-			const lines = buffer.split('\n');
-			buffer = lines.pop() ?? '';
-
-			for (const line of lines) {
-				if (!line.startsWith('data: ')) continue;
-				const raw = line.slice(6);
-
-				let data: Record<string, unknown>;
-				try {
-					data = JSON.parse(raw) as Record<string, unknown>;
-				} catch {
-					continue;
-				}
-
-				if (data.done) continue;
-
-				if (typeof data.text === 'string') {
-					ensureMsg();
-					assistantMsg.content += data.text;
-					scrollToBottom();
-				}
-
-				if (typeof data.thinking === 'string') {
-					ensureMsg();
-					assistantMsg.thinking = (assistantMsg.thinking ?? '') + data.thinking;
-				}
-
-				if (typeof data.codeDelta === 'string') {
-					emit('codeDelta', data.codeDelta);
-				}
-
-				if (typeof data.code === 'string') {
-					emit('codeUpdated');
-				}
-
-				if (data.configUpdated !== undefined || data.toolUpdated !== undefined) {
-					builderMutated = true;
-					emit('configUpdated');
-				}
-
-				if (data.toolCall && typeof data.toolCall === 'object') {
-					ensureMsg();
-					if (assistantMsg.content && !assistantMsg.content.endsWith('\n')) {
-						assistantMsg.content += '\n';
-					}
-					const tc = data.toolCall as { tool: string; input?: unknown };
-					assistantMsg.toolCalls = assistantMsg.toolCalls ?? [];
-					assistantMsg.toolCalls.push({ tool: tc.tool, input: tc.input });
-				}
-
-				if (data.toolResult && typeof data.toolResult === 'object') {
-					const tr = data.toolResult as { tool: string; output?: unknown };
-					const existing = assistantMsg.toolCalls?.find(
-						(t) => t.tool === tr.tool && t.output === undefined,
-					);
-					if (existing) {
-						existing.output = tr.output;
-					}
-					if (assistantMsg.content && !assistantMsg.content.endsWith('\n')) {
-						assistantMsg.content += '\n';
-					}
-				}
-
-				if (typeof data.error === 'string') {
-					ensureMsg();
-					assistantMsg.content += `\n\nError: ${data.error}`;
-					assistantMsg.status = 'error';
-				}
-			}
-		}
-
-		assistantMsg.status = 'success';
-	} catch (e) {
-		if (e instanceof DOMException && e.name === 'AbortError') {
-			assistantMsg.status = 'success';
-		} else {
-			if (!msgAdded) {
-				messages.value.push(assistantMsg);
-			}
-			assistantMsg.content = `Error: ${e instanceof Error ? e.message : 'Unknown error'}`;
-			assistantMsg.status = 'error';
-		}
-	} finally {
-		abortController.value = null;
-		isStreaming.value = false;
-		scrollToBottom();
-		if (endpoint === 'build' && builderMutated) {
-			emit('configUpdated');
-		}
-	}
-}
-
-async function sendMessage() {
-	const text = inputText.value.trim();
-	if (!text || isStreaming.value) return;
-
-	messages.value.push({
-		id: crypto.randomUUID(),
-		role: 'user',
-		content: text,
-		status: 'success',
-	});
+async function onSubmit() {
+	const text = inputText.value;
 	inputText.value = '';
-	scrollToBottom();
-
-	await streamFromEndpoint(props.endpoint, text);
+	await sendMessage(text);
 }
 
 function sendMessageFromOutside(message: string) {
 	inputText.value = message;
-	void sendMessage();
-}
-
-function stopGenerating() {
-	abortController.value?.abort();
-}
-
-function onSubmit() {
-	void sendMessage();
+	void onSubmit();
 }
 
 defineExpose({ sendMessageFromOutside });
@@ -346,103 +91,9 @@ onMounted(() => {
 			</button>
 		</div>
 
-		<!-- Messages -->
-		<div ref="scrollRef" :class="$style.messages">
-			<div v-if="messages.length === 0 && !isStreaming" :class="$style.emptyState">
-				<N8nIcon :icon="endpoint === 'build' ? 'wand-sparkles' : 'message-square'" :size="32" />
-				<N8nText tag="p" bold>
-					{{ endpoint === 'build' ? 'Build your agent' : 'Chat with your agent' }}
-				</N8nText>
-				<N8nText size="small" color="text-light">
-					{{
-						endpoint === 'build'
-							? 'Describe what you want your agent to do'
-							: 'Send a message to start a conversation'
-					}}
-				</N8nText>
-			</div>
-			<template v-else>
-				<div
-					v-for="msg in messages"
-					:key="msg.id"
-					:class="[$style.message, msg.role === 'user' ? $style.user : $style.assistant]"
-				>
-					<!-- Avatar -->
-					<div :class="$style.avatar">
-						<N8nIcon v-if="msg.role === 'user'" icon="user" width="20" height="20" />
-						<N8nIcon v-else icon="robot" width="20" height="20" />
-					</div>
+		<AgentChatEmptyState v-if="messages.length === 0 && !isStreaming" :endpoint="endpoint" />
+		<AgentChatMessageList v-else :messages="messages" :messaging-state="messagingState" />
 
-					<!-- Content -->
-					<div :class="$style.content">
-						<!-- Thinking -->
-						<details v-if="msg.thinking" :class="$style.thinkingBlock">
-							<summary :class="$style.thinkingSummary">
-								<N8nIcon icon="brain" :size="12" />
-								Thinking...
-							</summary>
-							<div :class="$style.thinkingContent">{{ msg.thinking }}</div>
-						</details>
-
-						<!-- Tool calls -->
-						<div v-if="msg.toolCalls?.length" :class="$style.toolCalls">
-							<div v-for="(tc, i) in msg.toolCalls" :key="i" :class="$style.toolCall">
-								<N8nIcon icon="wrench" :size="12" />
-								<span :class="$style.toolName">{{ tc.tool }}</span>
-								<N8nIcon
-									v-if="tc.output !== undefined"
-									icon="check"
-									:size="12"
-									:class="$style.toolDone"
-								/>
-								<ChatTypingIndicator v-else />
-							</div>
-						</div>
-
-						<!-- User message -->
-						<div v-if="msg.role === 'user'" :class="[$style.chatMessage, $style.chatMessageUser]">
-							{{ msg.content }}
-						</div>
-
-						<!-- Assistant message -->
-						<div
-							v-else-if="msg.content"
-							:class="[$style.chatMessage, { [$style.chatMessageError]: msg.status === 'error' }]"
-						>
-							<div :class="$style.markdownContent">
-								<ChatMarkdownChunk
-									:source="{ type: 'text', content: msg.content }"
-									@open-artifact="() => {}"
-								/>
-							</div>
-						</div>
-
-						<!-- Typing indicator for assistant still streaming with no content -->
-						<ChatTypingIndicator
-							v-if="
-								msg.role === 'assistant' &&
-								msg.status === 'streaming' &&
-								!msg.content &&
-								!msg.toolCalls?.length
-							"
-							:class="$style.typingIndicator"
-						/>
-					</div>
-				</div>
-
-				<!-- Waiting for first chunk indicator -->
-				<div v-if="messagingState === 'waitingFirstChunk'" :class="$style.message">
-					<div :class="$style.avatar">
-						<N8nIcon icon="robot" width="20" height="20" />
-					</div>
-					<div :class="$style.content">
-						<ChatTypingIndicator :class="$style.typingIndicator" />
-					</div>
-				</div>
-			</template>
-		</div>
-
-		<!-- Input -->
 		<div :class="$style.inputArea">
 			<ChatInputBase
 				v-model="inputText"
@@ -498,146 +149,6 @@ onMounted(() => {
 .clearBtn:hover {
 	background-color: var(--color--foreground--tint-1);
 	color: var(--color--danger);
-}
-
-/* Messages area — matches ChatHub layout */
-.messages {
-	flex: 1;
-	min-height: 0;
-	overflow-y: auto;
-	padding: var(--spacing--lg) var(--spacing--xl);
-	display: flex;
-	flex-direction: column;
-	gap: var(--spacing--lg);
-}
-
-.emptyState {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	height: 100%;
-	gap: var(--spacing--3xs);
-	color: var(--color--text--tint-2);
-}
-
-/* Message layout — mirrors ChatMessage.vue styles */
-.message {
-	position: relative;
-	padding-left: 40px;
-}
-
-.avatar {
-	position: absolute;
-	left: 0;
-	top: 0;
-	display: grid;
-	place-items: center;
-	width: 28px;
-	height: 28px;
-	border-radius: 50%;
-	background: var(--color--background);
-	color: var(--color--text--tint-1);
-}
-
-.content {
-	display: flex;
-	flex-direction: column;
-	align-items: stretch;
-}
-
-.chatMessage {
-	overflow-wrap: break-word;
-	font-size: var(--font-size--md);
-	line-height: var(--line-height--xl);
-}
-
-.chatMessageUser {
-	padding: var(--spacing--2xs) var(--spacing--sm);
-	border-radius: var(--radius--xl);
-	background-color: var(--color--background);
-	white-space: pre-wrap;
-	width: fit-content;
-	max-width: 100%;
-}
-
-.chatMessageError {
-	padding: var(--spacing--xs) var(--spacing--sm);
-	border-radius: var(--radius--lg);
-	background-color: var(--color--danger--tint-4);
-	border: var(--border-width) var(--border-style) var(--color--danger--tint-3);
-	color: var(--color--danger);
-}
-
-.markdownContent {
-	color: var(--color--text--shade-1);
-	font-size: var(--font-size--md);
-	line-height: var(--line-height--xl);
-
-	> *:last-child > *:last-child {
-		margin-bottom: 0;
-	}
-
-	> *:first-child > *:first-child {
-		margin-top: 0;
-	}
-}
-
-/* Thinking block */
-.thinkingBlock {
-	margin-bottom: var(--spacing--2xs);
-	font-size: var(--font-size--2xs);
-}
-
-.thinkingSummary {
-	cursor: pointer;
-	color: var(--color--text--tint-2);
-	font-style: italic;
-	display: flex;
-	align-items: center;
-	gap: var(--spacing--4xs);
-}
-
-.thinkingContent {
-	margin: var(--spacing--4xs) 0 0;
-	white-space: pre-wrap;
-	font-family: inherit;
-	font-size: var(--font-size--2xs);
-	color: var(--color--text--tint-1);
-	max-height: 150px;
-	overflow-y: auto;
-}
-
-/* Tool calls */
-.toolCalls {
-	display: flex;
-	flex-direction: column;
-	gap: var(--spacing--4xs);
-	margin-bottom: var(--spacing--2xs);
-}
-
-.toolCall {
-	display: inline-flex;
-	align-items: center;
-	gap: var(--spacing--4xs);
-	font-size: var(--font-size--2xs);
-	color: var(--color--text--tint-1);
-	padding: var(--spacing--4xs) var(--spacing--2xs);
-	background-color: var(--color--foreground--tint-2);
-	border-radius: var(--radius);
-	width: fit-content;
-}
-
-.toolName {
-	font-family: monospace;
-}
-
-.toolDone {
-	color: var(--color--success);
-}
-
-.typingIndicator {
-	margin-top: var(--spacing--xs);
 }
 
 .inputArea {
