+++ packages/frontend/editor-ui/src/features/agents/components/AgentChatToolSteps.vue
@@ -1,0 +1,85 @@
+<script setup lang="ts">
+import { N8nIcon } from '@n8n/design-system';
+import type { ToolCall } from '../composables/agentChatMessages';
+
+defineProps<{
+	toolCalls: ToolCall[];
+}>();
+</script>
+
+<template>
+	<ol :class="$style.toolSteps">
+		<li v-for="(tc, i) in toolCalls" :key="i" :class="$style.toolStep">
+			<div :class="$style.toolStepIndicator">
+				<N8nIcon
+					v-if="tc.output !== undefined"
+					icon="circle-check"
+					:size="14"
+					:class="$style.toolStepDone"
+				/>
+				<N8nIcon
+					v-else
+					icon="loader-circle"
+					:size="14"
+					:spin="true"
+					:class="$style.toolStepLoading"
+				/>
+			</div>
+			<span :class="$style.toolStepLabel">{{ tc.tool }}</span>
+		</li>
+	</ol>
+</template>
+
+<style module>
+.toolSteps {
+	list-style: none;
+	margin: 0 0 var(--spacing--2xs);
+	padding: 0;
+	display: flex;
+	flex-direction: column;
+	gap: var(--spacing--2xs);
+}
+
+.toolStep {
+	display: flex;
+	align-items: center;
+	gap: var(--spacing--xs);
+	position: relative;
+}
+
+.toolStepIndicator {
+	position: relative;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	width: 14px;
+	height: 14px;
+	flex-shrink: 0;
+	color: var(--color--text--tint-2);
+}
+
+.toolStep:not(:last-child) .toolStepIndicator::after {
+	content: '';
+	position: absolute;
+	top: calc(100% + 1px);
+	left: 50%;
+	width: 1px;
+	height: var(--spacing--2xs);
+	transform: translateX(-50%);
+	background-color: var(--color--foreground);
+}
+
+.toolStepDone {
+	color: var(--color--text--tint-2);
+}
+
+.toolStepLoading {
+	color: var(--color--text--tint-2);
+}
+
+.toolStepLabel {
+	font-family: monospace;
+	font-size: var(--font-size--2xs);
+	color: var(--color--text--tint-1);
+}
+</style>
